refactor: Reorganize "helpers" to be less generic
Align with Gutenberg project practices. Avoid generic filenames that
fail to describe intent or focus.
